feat: implement get payout channels on presentation layer

+++ app/src/main/java/com/prayatna/lookiesapp/data/remote/dto/PayoutChannelDto.kt
@@ -18,8 +18,8 @@ data class PayoutChannelDto(
 
 @Serializable
 data class AmountLimitsDto(
-    val minimum: Long,
-    val maximum: Long,
+    val minimum: Double,
+    val maximum: Double,
     @SerialName("minimum_increment")
-    val minimumIncrement: Long
+    val minimumIncrement: Double
 )
